install: add install script covering OS and Python dependencies
This commit is a preparation for a commit later in the patch series that
will remove the code in SConstruct that deals with installing Ubuntu
packages as well as Python packages. It also serves as a preparation for
start using git submodules to keep track of intra-dependencies like
HEANN, googletest, NTL and onnx.

The script is self provides an optional parameter "-v" that when enable
will install all the necessary Python packages into a janncy-local
folder called .pyvenv. By doing so it's possible to avoid potential
dependency issue that could show up on a system when different Python
scripts requires different versions of Python packages.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>