benchmark: adding a simple benchmark class
class BenchMark will make time measurements withing the local scope.
When instantiated it will record the current time and when it's no
longer in scope and the destructor is being called it'll record the
current time again and then it'll report the total runtime for the
scope.

If the object is instantiated at the top of the function then it'll be
the runtime for the entire function. If scope needs to be narrowed down
one can limit the scope, for example.

void foo() {
	str s;
	do_something_with(s);
	{
		BenchMark b(__func__);
		do_something_else_with(s);
	}
}

In the above example the runtime for 'do_something_with(s)' will not be
accounted.

The drawbacks with this way of making benchmarks is that is uses wall
clock time and not cycle instructions. Therefore the results will be
less reliable. I.e., different computers will give different results and
if the load would increase on the computer running this benchmark it
would also affect the results. This type of benchmark is best suited
when running it on the same computer, with no other unexpected CPU
demanding jobs.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>